+++ packages/core/src/RenderingEngine/helpers/cpuFallback/rendering/generateLut.ts
@@ -39,21 +39,43 @@ export default function (
   const mlutfn = getModalityLut(image.slope, image.intercept, modalityLUT);
   const vlutfn = getVOILUT(windowWidth, windowCenter, voiLUT);
 
-  if (invert === true) {
-    for (
-      let storedValue = minPixelValue;
-      storedValue <= maxPixelValue;
-      storedValue++
-    ) {
-      lut[storedValue + -offset] = 255 - vlutfn(mlutfn(storedValue));
+  if (image.isPreScaled) {
+    // if the image is already preScaled, it means that the slop and the intercept
+    // are applied and there is no need for a modalityLut
+    if (invert === true) {
+      for (
+        let storedValue = minPixelValue;
+        storedValue <= maxPixelValue;
+        storedValue++
+      ) {
+        lut[storedValue + -offset] = 255 - vlutfn(storedValue);
+      }
+    } else {
+      for (
+        let storedValue = minPixelValue;
+        storedValue <= maxPixelValue;
+        storedValue++
+      ) {
+        lut[storedValue + -offset] = vlutfn(storedValue);
+      }
     }
   } else {
-    for (
-      let storedValue = minPixelValue;
-      storedValue <= maxPixelValue;
-      storedValue++
-    ) {
-      lut[storedValue + -offset] = vlutfn(mlutfn(storedValue));
+    if (invert === true) {
+      for (
+        let storedValue = minPixelValue;
+        storedValue <= maxPixelValue;
+        storedValue++
+      ) {
+        lut[storedValue + -offset] = 255 - vlutfn(mlutfn(storedValue));
+      }
+    } else {
+      for (
+        let storedValue = minPixelValue;
+        storedValue <= maxPixelValue;
+        storedValue++
+      ) {
+        lut[storedValue + -offset] = vlutfn(mlutfn(storedValue));
+      }
     }
   }