+++ packages/tools/src/tools/StackScrollToolMouseWheelTool.ts
@@ -34,6 +34,7 @@ export default class StackScrollMouseWheelTool extends BaseTool {
       volumeId = targetId.split('volumeId:')[1];
     }
 
-    scrollThroughStack(viewport, { direction, invert, volumeId });
+    const delta = direction * (invert ? -1 : 1);
+    scrollThroughStack(viewport, { delta, volumeId });
   }
 }