+++ packages/tools/src/utilities/stackScrollTool/index.ts
@@ -1,5 +1,3 @@
-import snapFocalPointToSlice from './snapFocalPointToSlice';
-import getSliceRange from './getSliceRange';
 import scrollThroughStack from './scrollThroughStack';
 
-export { snapFocalPointToSlice, getSliceRange, scrollThroughStack };
+export { scrollThroughStack };