+++ packages/core/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0](https://github.com/cornerstonejs/cornerstone3D-beta/compare/@cornerstonejs/core@0.9.0...@cornerstonejs/core@0.10.0) (2022-05-24)
+
+
+### Features
+
+* Add VOLUME_NEW_IMAGE event and Add jumpToSlice and default VOI for volume viewport ([#104](https://github.com/cornerstonejs/cornerstone3D-beta/issues/104)) ([d36a23a](https://github.com/cornerstonejs/cornerstone3D-beta/commit/d36a23a4eaf5bafcc8dddc0ab796065098df616a))
+
+
+
+
+
+# [0.9.0](https://github.com/cornerstonejs/cornerstone3D-beta/compare/@cornerstonejs/core@0.8.1...@cornerstonejs/core@0.9.0) (2022-05-24)
+
+
+### Features
+
+* Add Clipping planes for rendering ([#110](https://github.com/cornerstonejs/cornerstone3D-beta/issues/110)) ([1a6e4c7](https://github.com/cornerstonejs/cornerstone3D-beta/commit/1a6e4c742a3b89a88b46fd98d6cbeca5c95918aa))
+
+
+
+
+
 ## [0.8.1](https://github.com/cornerstonejs/cornerstone3D-beta/compare/@cornerstonejs/core@0.8.0...@cornerstonejs/core@0.8.1) (2022-05-11)